XYZ Coords are nearly working
The calculation doesn't crash now, but there is still an issue with the way that some of the initial calculations are done.

Also, initial calculation of XYZ is way off.

+++ firmware/lib/Rube/src/Winch.h
@@ -78,7 +78,7 @@ public:
 
     // Position Variables
 
-    double startLength = 0.0;
+    float startLength = 0.0;
     /**< Current position in distance from start
      * Up (shortening the line) is negative.
      */
