+++ firmware/lib/Rube/src/WinchDriver.h
@@ -77,7 +77,7 @@ public:
 
     //TODO: Should not be getting from estimate here and from tracker in Winch
     float getPosition(){ return pos_est; }
-    float getEncoderTurns(){ return enc_pos; }
+    float getEncoderTurns(){ return (float) enc_pos; }
     long getEncoderTicks(){ return enc->read(); }
     float getSpeed(){ return spd_est; }
